refactor(mediawiki): split search.go into similarity/topic siblings (deslop #59)
