+++ wiki/search.go
@@ -6,7 +6,6 @@ import (
 	"html"
 	"net/url"
 	"regexp"
-	"sort"
 	"strconv"
 	"strings"
 )
@@ -252,365 +251,4 @@ func (c *Client) SearchInFile(ctx context.Context, args SearchInFileArgs) (Searc
 
 // FindSimilarPages finds pages with similar content to the given page
 // scoredCandidate is the intermediate representation for similarity ranking.
-type scoredCandidate struct {
-	title     string
-	score     float64
-	terms     []string
-	isLinked  bool
-	linksBack bool
-}
-
-// collectSimilarityCandidates returns candidate page titles for the similarity
-// search, drawn either from a category or from a full-text search query.
-// Excludes the source page itself.
-func (c *Client) collectSimilarityCandidates(ctx context.Context, source string, category, searchQuery string) []string {
-	var candidates []string
-	if category != "" {
-		catResult, err := c.GetCategoryMembers(ctx, CategoryMembersArgs{Category: category, Limit: 100})
-		if err != nil {
-			return nil
-		}
-		for _, member := range catResult.Members {
-			if member.Title != source {
-				candidates = append(candidates, member.Title)
-			}
-		}
-		return candidates
-	}
-	searchResult, err := c.Search(ctx, SearchArgs{Query: searchQuery, Limit: 50})
-	if err != nil {
-		return nil
-	}
-	for _, hit := range searchResult.Results {
-		if hit.Title != source {
-			candidates = append(candidates, hit.Title)
-		}
-	}
-	return candidates
-}
-
-// loadOutgoingLinkSet returns the set of titles that the given page links to.
-// On error, returns an empty set so callers can use it as a "no links" sentinel.
-func (c *Client) loadOutgoingLinkSet(ctx context.Context, page string, limit int) map[string]bool {
-	set := make(map[string]bool)
-	links, err := c.getPageLinks(ctx, page, limit)
-	if err != nil {
-		return set
-	}
-	for _, link := range links {
-		set[link.Title] = true
-	}
-	return set
-}
-
-// candidateLinksBackTo reports whether the candidate page contains a link to
-// the source title.
-func (c *Client) candidateLinksBackTo(ctx context.Context, candidate, source string) bool {
-	candLinks, err := c.getPageLinks(ctx, candidate, 500)
-	if err != nil {
-		return false
-	}
-	for _, link := range candLinks {
-		if link.Title == source {
-			return true
-		}
-	}
-	return false
-}
-
-// scoreSimilarityCandidate computes the similarity score and link metadata for
-// one candidate page. Returns ok=false if the candidate fails to fetch or
-// scores below the threshold.
-func (c *Client) scoreSimilarityCandidate(ctx context.Context, candidate, source string, sourceTerms []string, sourceLinks map[string]bool, minScore float64) (scoredCandidate, bool) {
-	candContent, err := c.GetPage(ctx, GetPageArgs{Title: candidate})
-	if err != nil {
-		return scoredCandidate{}, false
-	}
-	candTerms := extractKeyTerms(candContent.Content)
-	similarity := calculateJaccardSimilarity(sourceTerms, candTerms)
-	if similarity < minScore {
-		return scoredCandidate{}, false
-	}
-	return scoredCandidate{
-		title:     candidate,
-		score:     similarity,
-		terms:     findCommonTerms(sourceTerms, candTerms, 10),
-		isLinked:  sourceLinks[candidate],
-		linksBack: c.candidateLinksBackTo(ctx, candidate, source),
-	}, true
-}
-
-// similarityRecommendation returns advice for the user given the link state
-// between the source and candidate pages.
-func similarityRecommendation(score float64, isLinked, linksBack bool, source, candidate string) string {
-	switch {
-	case score > 0.6 && !isLinked && !linksBack:
-		return "Possible duplicate - high similarity but no links between pages"
-	case !isLinked && !linksBack:
-		return "Consider cross-linking - related content but no links"
-	case isLinked && !linksBack:
-		return fmt.Sprintf("Add backlink from '%s' to '%s'", candidate, source)
-	case !isLinked && linksBack:
-		return fmt.Sprintf("Add link from '%s' to '%s'", source, candidate)
-	default:
-		return "Already cross-linked"
-	}
-}
-
-func (c *Client) FindSimilarPages(ctx context.Context, args FindSimilarPagesArgs) (FindSimilarPagesResult, error) {
-	if args.Page == "" {
-		return FindSimilarPagesResult{}, fmt.Errorf("page is required")
-	}
-	if err := c.EnsureLoggedIn(ctx); err != nil {
-		return FindSimilarPagesResult{}, err
-	}
-
-	limit := normalizeLimit(args.Limit, 10, 50)
-	minScore := args.MinScore
-	if minScore <= 0 {
-		minScore = 0.1
-	}
-	source := normalizePageTitle(args.Page)
-
-	sourceContent, err := c.GetPage(ctx, GetPageArgs{Title: source})
-	if err != nil {
-		return FindSimilarPagesResult{}, fmt.Errorf("failed to get source page: %w", err)
-	}
-	sourceTerms := extractKeyTerms(sourceContent.Content)
-	if len(sourceTerms) == 0 {
-		return FindSimilarPagesResult{
-			SourcePage:   source,
-			SimilarPages: []SimilarPage{},
-			Message:      "Source page has no significant terms for comparison",
-		}, nil
-	}
-
-	searchQuery := strings.Join(extractTopTerms(sourceContent.Content, 5), " ")
-	candidates := c.collectSimilarityCandidates(ctx, source, args.Category, searchQuery)
-	if len(candidates) == 0 {
-		return FindSimilarPagesResult{
-			SourcePage:   source,
-			SimilarPages: []SimilarPage{},
-			Message:      "No candidate pages found for comparison",
-		}, nil
-	}
-
-	sourceLinks := c.loadOutgoingLinkSet(ctx, source, 500)
-
-	scored := make([]scoredCandidate, 0)
-	for _, candidate := range candidates {
-		if ctx.Err() != nil {
-			break
-		}
-		if sc, ok := c.scoreSimilarityCandidate(ctx, candidate, source, sourceTerms, sourceLinks, minScore); ok {
-			scored = append(scored, sc)
-		}
-	}
-	sort.Slice(scored, func(i, j int) bool { return scored[i].score > scored[j].score })
-
-	similarPages := make([]SimilarPage, 0, limit)
-	for i, sp := range scored {
-		if i >= limit {
-			break
-		}
-		similarPages = append(similarPages, SimilarPage{
-			Title:           sp.title,
-			SimilarityScore: sp.score,
-			CommonTerms:     sp.terms,
-			IsLinked:        sp.isLinked,
-			LinksBack:       sp.linksBack,
-			Recommendation:  similarityRecommendation(sp.score, sp.isLinked, sp.linksBack, source, sp.title),
-		})
-	}
-
-	return FindSimilarPagesResult{
-		SourcePage:    source,
-		SimilarPages:  similarPages,
-		TotalCompared: len(candidates),
-	}, nil
-}
-
-// CompareTopic compares how a topic is described across multiple pages
 // pageValueRef tracks one value extracted near the topic, with the page it came from.
-type pageValueRef struct {
-	page  string
-	value string
-}
-
-// findPagesForTopic returns candidate page titles for topic comparison, drawn
-// from a category if specified, otherwise from a full-text search.
-func (c *Client) findPagesForTopic(ctx context.Context, topic, category string, searchLimit int) ([]string, error) {
-	var pageTitles []string
-	if category != "" {
-		catResult, err := c.GetCategoryMembers(ctx, CategoryMembersArgs{Category: category, Limit: 100})
-		if err != nil {
-			return pageTitles, nil //nolint:nilerr // category lookup failure falls through to "no candidates"
-		}
-		for _, member := range catResult.Members {
-			pageTitles = append(pageTitles, member.Title)
-		}
-		return pageTitles, nil
-	}
-	searchResult, err := c.Search(ctx, SearchArgs{Query: topic, Limit: searchLimit})
-	if err != nil {
-		return nil, fmt.Errorf("search failed: %w", err)
-	}
-	for _, hit := range searchResult.Results {
-		pageTitles = append(pageTitles, hit.Title)
-	}
-	return pageTitles, nil
-}
-
-// extractTopicValues collects values from the page that appear near a context
-// where the topic is mentioned. Caller-supplied contexts are the windows in
-// which co-occurrence is checked.
-func extractTopicValues(content, topic string, contexts []string, page string) map[string][]pageValueRef {
-	out := make(map[string][]pageValueRef)
-	topicLower := strings.ToLower(topic)
-	for _, v := range extractValues(content) {
-		valueCtxLower := strings.ToLower(v.Context)
-		for _, ctxStr := range contexts {
-			if strings.Contains(strings.ToLower(ctxStr), valueCtxLower) ||
-				strings.Contains(valueCtxLower, topicLower) {
-				out[v.Type] = append(out[v.Type], pageValueRef{page: page, value: v.Value})
-				break
-			}
-		}
-	}
-	return out
-}
-
-// analyzeTopicOnPage fetches the page, checks for the topic, and returns the
-// mention summary plus any near-topic values. Returns ok=false when the page
-// either fails to fetch or doesn't mention the topic.
-func (c *Client) analyzeTopicOnPage(ctx context.Context, title, topic string) (TopicMention, map[string][]pageValueRef, bool) {
-	page, err := c.GetPage(ctx, GetPageArgs{Title: title})
-	if err != nil {
-		return TopicMention{}, nil, false
-	}
-	contentLower := strings.ToLower(page.Content)
-	topicLower := strings.ToLower(topic)
-	if !strings.Contains(contentLower, topicLower) {
-		return TopicMention{}, nil, false
-	}
-	contexts := extractContextsForTerm(page.Content, topic, 3)
-	info, _ := c.GetPageInfo(ctx, PageInfoArgs{Title: title})
-	mention := TopicMention{
-		PageTitle:  title,
-		Mentions:   strings.Count(contentLower, topicLower),
-		Contexts:   contexts,
-		LastEdited: info.Touched,
-	}
-	return mention, extractTopicValues(page.Content, topic, contexts, title), true
-}
-
-// compareValuePair returns an Inconsistency when two page-value refs disagree
-// after normalization. The second return is false when there is nothing to report.
-func compareValuePair(valueType string, a, b pageValueRef) (Inconsistency, bool) {
-	if a.page == b.page {
-		return Inconsistency{}, false
-	}
-	v1 := normalizeValue(a.value)
-	v2 := normalizeValue(b.value)
-	if v1 == v2 || v1 == "" || v2 == "" {
-		return Inconsistency{}, false
-	}
-	return Inconsistency{
-		Type:        valueType,
-		Description: fmt.Sprintf("%s values differ", valueType),
-		PageA:       a.page,
-		PageB:       b.page,
-		ValueA:      a.value,
-		ValueB:      b.value,
-	}, true
-}
-
-// collectInconsistenciesForType returns all inconsistent value pairs within a
-// single value type's page-value refs.
-func collectInconsistenciesForType(valueType string, pageValues []pageValueRef) []Inconsistency {
-	var out []Inconsistency
-	for i := 0; i < len(pageValues)-1; i++ {
-		for j := i + 1; j < len(pageValues); j++ {
-			if inc, ok := compareValuePair(valueType, pageValues[i], pageValues[j]); ok {
-				out = append(out, inc)
-			}
-		}
-	}
-	return out
-}
-
-// detectValueInconsistencies pairs up cross-page values of the same type and
-// reports any that disagree once normalized.
-func detectValueInconsistencies(allValues map[string][]pageValueRef) []Inconsistency {
-	inconsistencies := make([]Inconsistency, 0)
-	for valueType, pageValues := range allValues {
-		if len(pageValues) < 2 {
-			continue
-		}
-		inconsistencies = append(inconsistencies, collectInconsistenciesForType(valueType, pageValues)...)
-	}
-	return inconsistencies
-}
-
-func (c *Client) CompareTopic(ctx context.Context, args CompareTopicArgs) (CompareTopicResult, error) {
-	if args.Topic == "" {
-		return CompareTopicResult{}, fmt.Errorf("topic is required")
-	}
-	if err := c.EnsureLoggedIn(ctx); err != nil {
-		return CompareTopicResult{}, err
-	}
-
-	limit := normalizeLimit(args.Limit, 20, 50)
-	pageTitles, err := c.findPagesForTopic(ctx, args.Topic, args.Category, limit*2)
-	if err != nil {
-		return CompareTopicResult{}, err
-	}
-	if len(pageTitles) == 0 {
-		return CompareTopicResult{
-			Topic:        args.Topic,
-			PageMentions: []TopicMention{},
-			Summary:      fmt.Sprintf("No pages found mentioning '%s'", args.Topic),
-		}, nil
-	}
-
-	mentions := make([]TopicMention, 0)
-	allValues := make(map[string][]pageValueRef)
-	for _, title := range pageTitles {
-		if len(mentions) >= limit || ctx.Err() != nil {
-			break
-		}
-		mention, values, ok := c.analyzeTopicOnPage(ctx, title, args.Topic)
-		if !ok {
-			continue
-		}
-		mentions = append(mentions, mention)
-		for valueType, refs := range values {
-			allValues[valueType] = append(allValues[valueType], refs...)
-		}
-	}
-
-	inconsistencies := detectValueInconsistencies(allValues)
-	summary := fmt.Sprintf("Found %d pages mentioning '%s'", len(mentions), args.Topic)
-	if len(inconsistencies) > 0 {
-		summary += fmt.Sprintf(". Detected %d potential inconsistencies", len(inconsistencies))
-	}
-
-	return CompareTopicResult{
-		Topic:           args.Topic,
-		PagesFound:      len(mentions),
-		PageMentions:    mentions,
-		Inconsistencies: inconsistencies,
-		Summary:         summary,
-	}, nil
-}
-
-// normalizeValue extracts the core numeric value for comparison
-func normalizeValue(value string) string {
-	// Extract just numbers for comparison
-	re := regexp.MustCompile(`\d+(?:\.\d+)?`)
-	matches := re.FindAllString(value, -1)
-	if len(matches) > 0 {
-		return matches[0]
-	}
-	return strings.TrimSpace(strings.ToLower(value))
-}
